Add velocity of SpaceShip to the final velocity of the Laser.
When launching something with a cannon or a laser gun, the projectile's
velocity increases relative to the launcher's velocity.
According to the theory of relativity, this velocity will not perfectly
add up with the velocity of the launcher nearing c, but in our case the
velocities are rather slow.
Therefore we can just add the velocities up.