Fixed Asteroid-spawns-in-front-of-your-ship bug.
What caused this bug was Box2D reusing the body objects and thus reusing
the position vector. Therefore when an Asteroid despawned and - for
example - a Laser was spawned, the Vector2 in the asteroid splitter was
changed and the new asteroids were spawned in front of the spaceship.

The bug was fixed by copying the Vector2 and then passing it to the
asteroid splitter, so the Vector2 would not be modified by any movements
of the Box2D body.