Added a meaning to MetalAsteroid and ExplosiveAsteroid.
This also includes MetalAsteroidProjectile, which is spawned when a
MetalAsteroid is destroyed. Those projectiles behave just like lasers,
but they only ever fly in a straight path and do 25 damage to whatever
they hit. They also propagate score tracking, so if a player hits a
MetalAsteroid, causing projectiles to spawn which hit and destroy
another asteroid, the player gets (twice the usual) score for this
heroic feat.